requirements.txt
# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Email processing
email-validator==2.1.0.post1

# Data processing
pandas==2.1.4
python-dateutil==2.8.2

# Environment management
python-dotenv==1.0.0

# Logging and utilities
coloredlogs==15.0.1
schedule==1.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1